extracted previous communications into its own component
